backend/requirements.txt
amqp==5.1.1
asgiref==3.3.1
astroid==2.3.1
astropy==5.2.1
autopep8==1.4.4
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.2
certifi==2020.12.5
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.0.7
cycler==0.11.0
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==3.1.8
django-auth-ldap==2.1.1
django-celery-results==2.2.0
django-cors-headers==3.3.0
django-filter==2.2.0
django-url-filter==0.3.15
djangorestframework==3.12.2
enum-compat==0.0.3
Flask==1.1.2
flower==2.0.1
fonttools==4.39.2
humanize==3.1.0
idna==2.10
importlib-metadata==3.7.2
importlib-resources==5.12.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.1.0
kombu==5.3.1
lazy-object-proxy==1.4.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.7.1
mccabe==0.6.1
numpy==1.21.1
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
plotly==5.14.1
prometheus-client==0.8.0
prompt-toolkit==3.0.39
psycopg2-binary==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyerfa==2.0.0.1
pyflakes==2.2.0
pylibmc==1.6.3
pyparsing==2.4.2
python-dateutil==2.8.0
python-ldap==3.4.0
pytz==2021.1
PyYAML==6.0
requests==2.31.0
scipy==1.5.4
Shapely==1.7.1
six==1.12.0
sqlparse==0.4.4
tenacity==8.2.3
tornado==6.3.3
tqdm==4.66.1
typed-ast==1.4.0
typing-extensions==3.7.4.3
tzdata==2023.3
urllib3==1.26.5
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.2.6
Werkzeug==0.16.0
wrapt==1.11.2
zipp==3.4.1
black==23.9.1
